refactor: Comprehensive security, performance, and code quality improvements
Security fixes:
- Add timing-safe comparison to cron endpoints (cleanup, weather)
- Add security headers (CSP, HSTS, X-Frame-Options, etc.) in next.config.mjs
- Strengthen password requirements to 12 chars with complexity validation
- Add x-request-id headers for request traceability in all API routes
- Remove console statements that could leak sensitive information

Code quality improvements:
- Split ride-comparison-form.tsx (1306→741 lines) into location-input.tsx,
  airport-selector.tsx, useLocationSuggestions.ts, useUserLocation.ts
- Split map.tsx (1382→151 lines) into map-markers.tsx, map-route.tsx,
  map-cluster.tsx, map-controls.tsx
- Split database.ts (908→69 lines) into database-routes.ts,
  database-logging.ts, database-pricing-stats.ts
- Remove 20+ console.log/warn/error statements from components and API routes

Performance optimizations:
- Add useMemo for services array, bestPrice, bestWaitTime calculations
- Add useCallback for all event handlers in ride-comparison-results.tsx
- Add bounded cache with FIFO eviction (10k max entries) to rate limiter
- Use immutable update patterns in rate limiter

New utilities:
- lib/service-mappings.ts: Centralized type-safe service type mapping
- lib/currency.ts: Price parsing and formatting utilities
