+++ components/ui/map.tsx
@@ -1,27 +1,33 @@
 'use client'
 
-import MapLibreGL, { type PopupOptions, type MarkerOptions } from 'maplibre-gl'
+import MapLibreGL from 'maplibre-gl'
 import 'maplibre-gl/dist/maplibre-gl.css'
 import { useTheme } from 'next-themes'
 import {
   createContext,
-  useCallback,
   useContext,
   useEffect,
-  useId,
   useMemo,
   useRef,
   useState,
   type ReactNode,
 } from 'react'
-import { createPortal } from 'react-dom'
-import { X, Minus, Plus, Locate, Maximize, Loader2 } from 'lucide-react'
 
-import { cn } from '@/lib/utils'
+// Re-export from extracted modules for backward compatibility
+export {
+  MapMarker,
+  MarkerContent,
+  MarkerPopup,
+  MarkerTooltip,
+  MarkerLabel,
+  MapPopup,
+} from './map-markers'
 
-// Exported constants for layer ID prefixes (used by E2E tests to decouple from internal naming)
-export const ROUTE_LAYER_PREFIX = 'route-layer-'
-export const ROUTE_SOURCE_PREFIX = 'route-source-'
+export { MapRoute, ROUTE_LAYER_PREFIX, ROUTE_SOURCE_PREFIX } from './map-route'
+
+export { MapClusterLayer } from './map-cluster'
+
+export { MapControls } from './map-controls'
 
 type MapContextValue = {
   map: MapLibreGL.Map | null
@@ -142,1241 +148,4 @@ function Map({ children, styles, ...props }: MapProps) {
   )
 }
 
-type MarkerContextValue = {
-  markerRef: React.RefObject<MapLibreGL.Marker | null>
-  markerElementRef: React.RefObject<HTMLDivElement | null>
-  map: MapLibreGL.Map | null
-  isReady: boolean
-}
-
-const MarkerContext = createContext<MarkerContextValue | null>(null)
-
-function useMarkerContext() {
-  const context = useContext(MarkerContext)
-  if (!context) {
-    throw new Error('Marker components must be used within MapMarker')
-  }
-  return context
-}
-
-type MapMarkerProps = {
-  /** Longitude coordinate for marker position */
-  longitude: number
-  /** Latitude coordinate for marker position */
-  latitude: number
-  /** Marker subcomponents (MarkerContent, MarkerPopup, MarkerTooltip, MarkerLabel) */
-  children: ReactNode
-  /** Callback when marker is clicked */
-  onClick?: (e: MouseEvent) => void
-  /** Callback when mouse enters marker */
-  onMouseEnter?: (e: MouseEvent) => void
-  /** Callback when mouse leaves marker */
-  onMouseLeave?: (e: MouseEvent) => void
-  /** Callback when marker drag starts (requires draggable: true) */
-  onDragStart?: (lngLat: { lng: number; lat: number }) => void
-  /** Callback during marker drag (requires draggable: true) */
-  onDrag?: (lngLat: { lng: number; lat: number }) => void
-  /** Callback when marker drag ends (requires draggable: true) */
-  onDragEnd?: (lngLat: { lng: number; lat: number }) => void
-} & Omit<MarkerOptions, 'element'>
-
-function MapMarker({
-  longitude,
-  latitude,
-  children,
-  onClick,
-  onMouseEnter,
-  onMouseLeave,
-  onDragStart,
-  onDrag,
-  onDragEnd,
-  draggable = false,
-  ...markerOptions
-}: MapMarkerProps) {
-  const { map, isLoaded } = useMap()
-  const markerRef = useRef<MapLibreGL.Marker | null>(null)
-  const markerElementRef = useRef<HTMLDivElement | null>(null)
-  const [isReady, setIsReady] = useState(false)
-  const markerOptionsRef = useRef(markerOptions)
-
-  useEffect(() => {
-    if (!isLoaded || !map) return
-
-    const container = document.createElement('div')
-    markerElementRef.current = container
-
-    const marker = new MapLibreGL.Marker({
-      ...markerOptions,
-      element: container,
-      draggable,
-    })
-      .setLngLat([longitude, latitude])
-      .addTo(map)
-
-    markerRef.current = marker
-
-    const handleClick = (e: MouseEvent) => onClick?.(e)
-    const handleMouseEnter = (e: MouseEvent) => onMouseEnter?.(e)
-    const handleMouseLeave = (e: MouseEvent) => onMouseLeave?.(e)
-
-    container.addEventListener('click', handleClick)
-    container.addEventListener('mouseenter', handleMouseEnter)
-    container.addEventListener('mouseleave', handleMouseLeave)
-
-    const handleDragStart = () => {
-      const lngLat = marker.getLngLat()
-      onDragStart?.({ lng: lngLat.lng, lat: lngLat.lat })
-    }
-    const handleDrag = () => {
-      const lngLat = marker.getLngLat()
-      onDrag?.({ lng: lngLat.lng, lat: lngLat.lat })
-    }
-    const handleDragEnd = () => {
-      const lngLat = marker.getLngLat()
-      onDragEnd?.({ lng: lngLat.lng, lat: lngLat.lat })
-    }
-
-    marker.on('dragstart', handleDragStart)
-    marker.on('drag', handleDrag)
-    marker.on('dragend', handleDragEnd)
-
-    setIsReady(true)
-
-    return () => {
-      container.removeEventListener('click', handleClick)
-      container.removeEventListener('mouseenter', handleMouseEnter)
-      container.removeEventListener('mouseleave', handleMouseLeave)
-
-      marker.off('dragstart', handleDragStart)
-      marker.off('drag', handleDrag)
-      marker.off('dragend', handleDragEnd)
-
-      marker.remove()
-      markerRef.current = null
-      markerElementRef.current = null
-      setIsReady(false)
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [map, isLoaded])
-
-  useEffect(() => {
-    markerRef.current?.setLngLat([longitude, latitude])
-  }, [longitude, latitude])
-
-  useEffect(() => {
-    markerRef.current?.setDraggable(draggable)
-  }, [draggable])
-
-  useEffect(() => {
-    if (!markerRef.current) return
-    const prev = markerOptionsRef.current
-
-    if (prev.offset !== markerOptions.offset) {
-      markerRef.current.setOffset(markerOptions.offset ?? [0, 0])
-    }
-    if (prev.rotation !== markerOptions.rotation) {
-      markerRef.current.setRotation(markerOptions.rotation ?? 0)
-    }
-    if (prev.rotationAlignment !== markerOptions.rotationAlignment) {
-      markerRef.current.setRotationAlignment(markerOptions.rotationAlignment ?? 'auto')
-    }
-    if (prev.pitchAlignment !== markerOptions.pitchAlignment) {
-      markerRef.current.setPitchAlignment(markerOptions.pitchAlignment ?? 'auto')
-    }
-
-    markerOptionsRef.current = markerOptions
-  }, [markerOptions])
-
-  return (
-    <MarkerContext.Provider value={{ markerRef, markerElementRef, map, isReady }}>
-      {children}
-    </MarkerContext.Provider>
-  )
-}
-
-type MarkerContentProps = {
-  /** Custom marker content. Defaults to a blue dot if not provided */
-  children?: ReactNode
-  /** Additional CSS classes for the marker container */
-  className?: string
-}
-
-function MarkerContent({ children, className }: MarkerContentProps) {
-  const { markerElementRef, isReady } = useMarkerContext()
-
-  if (!isReady || !markerElementRef.current) return null
-
-  return createPortal(
-    <div className={cn('relative cursor-pointer', className)}>
-      {children || <DefaultMarkerIcon />}
-    </div>,
-    markerElementRef.current
-  )
-}
-
-function DefaultMarkerIcon() {
-  return (
-    <div className="relative h-4 w-4 rounded-full border-2 border-white bg-blue-500 shadow-lg" />
-  )
-}
-
-type MarkerPopupProps = {
-  /** Popup content */
-  children: ReactNode
-  /** Additional CSS classes for the popup container */
-  className?: string
-  /** Show a close button in the popup (default: false) */
-  closeButton?: boolean
-} & Omit<PopupOptions, 'className' | 'closeButton'>
-
-function MarkerPopup({
-  children,
-  className,
-  closeButton = false,
-  ...popupOptions
-}: MarkerPopupProps) {
-  const { markerRef, isReady } = useMarkerContext()
-  const containerRef = useRef<HTMLDivElement | null>(null)
-  const popupRef = useRef<MapLibreGL.Popup | null>(null)
-  const [mounted, setMounted] = useState(false)
-  const popupOptionsRef = useRef(popupOptions)
-
-  useEffect(() => {
-    if (!isReady || !markerRef.current) return
-
-    const container = document.createElement('div')
-    containerRef.current = container
-
-    const popup = new MapLibreGL.Popup({
-      offset: 16,
-      ...popupOptions,
-      closeButton: false,
-    })
-      .setMaxWidth('none')
-      .setDOMContent(container)
-
-    popupRef.current = popup
-    markerRef.current.setPopup(popup)
-    setMounted(true)
-
-    return () => {
-      popup.remove()
-      popupRef.current = null
-      containerRef.current = null
-      setMounted(false)
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [isReady])
-
-  useEffect(() => {
-    if (!popupRef.current) return
-    const prev = popupOptionsRef.current
-
-    if (prev.offset !== popupOptions.offset) {
-      popupRef.current.setOffset(popupOptions.offset ?? 16)
-    }
-    if (prev.maxWidth !== popupOptions.maxWidth && popupOptions.maxWidth) {
-      popupRef.current.setMaxWidth(popupOptions.maxWidth ?? 'none')
-    }
-
-    popupOptionsRef.current = popupOptions
-  }, [popupOptions])
-
-  const handleClose = () => popupRef.current?.remove()
-
-  if (!mounted || !containerRef.current) return null
-
-  return createPortal(
-    <div
-      className={cn(
-        'relative rounded-md border bg-popover p-3 text-popover-foreground shadow-md animate-in fade-in-0 zoom-in-95',
-        className
-      )}
-    >
-      {closeButton && (
-        <button
-          type="button"
-          onClick={handleClose}
-          className="absolute top-1 right-1 z-10 rounded-sm opacity-70 ring-offset-background transition-opacity hover:opacity-100 focus:outline-none focus:ring-2 focus:ring-ring focus:ring-offset-2"
-          aria-label="Close popup"
-        >
-          <X className="h-4 w-4" />
-          <span className="sr-only">Close</span>
-        </button>
-      )}
-      {children}
-    </div>,
-    containerRef.current
-  )
-}
-
-type MarkerTooltipProps = {
-  /** Tooltip content */
-  children: ReactNode
-  /** Additional CSS classes for the tooltip container */
-  className?: string
-} & Omit<PopupOptions, 'className' | 'closeButton' | 'closeOnClick'>
-
-function MarkerTooltip({ children, className, ...popupOptions }: MarkerTooltipProps) {
-  const { markerRef, markerElementRef, map, isReady } = useMarkerContext()
-  const containerRef = useRef<HTMLDivElement | null>(null)
-  const popupRef = useRef<MapLibreGL.Popup | null>(null)
-  const [mounted, setMounted] = useState(false)
-  const popupOptionsRef = useRef(popupOptions)
-
-  useEffect(() => {
-    if (!isReady || !markerRef.current || !markerElementRef.current || !map) return
-
-    const container = document.createElement('div')
-    containerRef.current = container
-
-    const popup = new MapLibreGL.Popup({
-      offset: 16,
-      ...popupOptions,
-      closeOnClick: true,
-      closeButton: false,
-    })
-      .setMaxWidth('none')
-      .setDOMContent(container)
-
-    popupRef.current = popup
-
-    const markerElement = markerElementRef.current
-    const marker = markerRef.current
-
-    const handleMouseEnter = () => {
-      popup.setLngLat(marker.getLngLat()).addTo(map)
-    }
-    const handleMouseLeave = () => popup.remove()
-
-    markerElement.addEventListener('mouseenter', handleMouseEnter)
-    markerElement.addEventListener('mouseleave', handleMouseLeave)
-    setMounted(true)
-
-    return () => {
-      markerElement.removeEventListener('mouseenter', handleMouseEnter)
-      markerElement.removeEventListener('mouseleave', handleMouseLeave)
-      popup.remove()
-      popupRef.current = null
-      containerRef.current = null
-      setMounted(false)
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [isReady, map])
-
-  useEffect(() => {
-    if (!popupRef.current) return
-    const prev = popupOptionsRef.current
-
-    if (prev.offset !== popupOptions.offset) {
-      popupRef.current.setOffset(popupOptions.offset ?? 16)
-    }
-    if (prev.maxWidth !== popupOptions.maxWidth && popupOptions.maxWidth) {
-      popupRef.current.setMaxWidth(popupOptions.maxWidth ?? 'none')
-    }
-
-    popupOptionsRef.current = popupOptions
-  }, [popupOptions])
-
-  if (!mounted || !containerRef.current) return null
-
-  return createPortal(
-    <div
-      className={cn(
-        'rounded-md bg-foreground px-2 py-1 text-xs text-background shadow-md animate-in fade-in-0 zoom-in-95',
-        className
-      )}
-    >
-      {children}
-    </div>,
-    containerRef.current
-  )
-}
-
-type MarkerLabelProps = {
-  /** Label text content */
-  children: ReactNode
-  /** Additional CSS classes for the label */
-  className?: string
-  /** Position of the label relative to the marker (default: "top") */
-  position?: 'top' | 'bottom'
-}
-
-function MarkerLabel({ children, className, position = 'top' }: MarkerLabelProps) {
-  const positionClasses = {
-    top: 'bottom-full mb-1',
-    bottom: 'top-full mt-1',
-  }
-
-  return (
-    <div
-      className={cn(
-        'absolute left-1/2 -translate-x-1/2 whitespace-nowrap',
-        'text-[10px] font-medium text-foreground',
-        positionClasses[position],
-        className
-      )}
-    >
-      {children}
-    </div>
-  )
-}
-
-type MapControlsProps = {
-  /** Position of the controls on the map (default: "bottom-right") */
-  position?: 'top-left' | 'top-right' | 'bottom-left' | 'bottom-right'
-  /** Show zoom in/out buttons (default: true) */
-  showZoom?: boolean
-  /** Show compass button to reset bearing (default: false) */
-  showCompass?: boolean
-  /** Show locate button to find user's location (default: false) */
-  showLocate?: boolean
-  /** Show fullscreen toggle button (default: false) */
-  showFullscreen?: boolean
-  /** Additional CSS classes for the controls container */
-  className?: string
-  /** Callback with user coordinates when located */
-  onLocate?: (coords: { longitude: number; latitude: number }) => void
-}
-
-const positionClasses = {
-  'top-left': 'top-2 left-2',
-  'top-right': 'top-2 right-2',
-  'bottom-left': 'bottom-2 left-2',
-  'bottom-right': 'bottom-10 right-2',
-}
-
-function ControlGroup({ children }: { children: React.ReactNode }) {
-  return (
-    <div className="flex flex-col rounded-md border border-border bg-background shadow-sm overflow-hidden [&>button:not(:last-child)]:border-b [&>button:not(:last-child)]:border-border">
-      {children}
-    </div>
-  )
-}
-
-function ControlButton({
-  onClick,
-  label,
-  children,
-  disabled = false,
-}: {
-  onClick: () => void
-  label: string
-  children: React.ReactNode
-  disabled?: boolean
-}) {
-  return (
-    <button
-      onClick={onClick}
-      aria-label={label}
-      type="button"
-      className={cn(
-        'flex items-center justify-center size-8 hover:bg-accent dark:hover:bg-accent/40 transition-colors',
-        disabled && 'opacity-50 pointer-events-none cursor-not-allowed'
-      )}
-      disabled={disabled}
-    >
-      {children}
-    </button>
-  )
-}
-
-function MapControls({
-  position = 'bottom-right',
-  showZoom = true,
-  showCompass = false,
-  showLocate = false,
-  showFullscreen = false,
-  className,
-  onLocate,
-}: MapControlsProps) {
-  const { map, isLoaded } = useMap()
-  const [waitingForLocation, setWaitingForLocation] = useState(false)
-
-  const handleZoomIn = useCallback(() => {
-    map?.zoomTo(map.getZoom() + 1, { duration: 300 })
-  }, [map])
-
-  const handleZoomOut = useCallback(() => {
-    map?.zoomTo(map.getZoom() - 1, { duration: 300 })
-  }, [map])
-
-  const handleResetBearing = useCallback(() => {
-    map?.resetNorthPitch({ duration: 300 })
-  }, [map])
-
-  const handleLocate = useCallback(() => {
-    setWaitingForLocation(true)
-    if ('geolocation' in navigator) {
-      navigator.geolocation.getCurrentPosition(
-        pos => {
-          const coords = {
-            longitude: pos.coords.longitude,
-            latitude: pos.coords.latitude,
-          }
-          map?.flyTo({
-            center: [coords.longitude, coords.latitude],
-            zoom: 14,
-            duration: 1500,
-          })
-          onLocate?.(coords)
-          setWaitingForLocation(false)
-        },
-        error => {
-          console.error('Error getting location:', error)
-          setWaitingForLocation(false)
-        }
-      )
-    }
-  }, [map, onLocate])
-
-  const handleFullscreen = useCallback(() => {
-    const container = map?.getContainer()
-    if (!container) return
-    if (document.fullscreenElement) {
-      document.exitFullscreen()
-    } else {
-      container.requestFullscreen()
-    }
-  }, [map])
-
-  if (!isLoaded) return null
-
-  return (
-    <div
-      className={cn('absolute z-10 flex flex-col gap-1.5', positionClasses[position], className)}
-    >
-      {showZoom && (
-        <ControlGroup>
-          <ControlButton onClick={handleZoomIn} label="Zoom in">
-            <Plus className="size-4" />
-          </ControlButton>
-          <ControlButton onClick={handleZoomOut} label="Zoom out">
-            <Minus className="size-4" />
-          </ControlButton>
-        </ControlGroup>
-      )}
-      {showCompass && (
-        <ControlGroup>
-          <CompassButton onClick={handleResetBearing} />
-        </ControlGroup>
-      )}
-      {showLocate && (
-        <ControlGroup>
-          <ControlButton
-            onClick={handleLocate}
-            label="Find my location"
-            disabled={waitingForLocation}
-          >
-            {waitingForLocation ? (
-              <Loader2 className="size-4 animate-spin" />
-            ) : (
-              <Locate className="size-4" />
-            )}
-          </ControlButton>
-        </ControlGroup>
-      )}
-      {showFullscreen && (
-        <ControlGroup>
-          <ControlButton onClick={handleFullscreen} label="Toggle fullscreen">
-            <Maximize className="size-4" />
-          </ControlButton>
-        </ControlGroup>
-      )}
-    </div>
-  )
-}
-
-function CompassButton({ onClick }: { onClick: () => void }) {
-  const { isLoaded, map } = useMap()
-  const compassRef = useRef<SVGSVGElement>(null)
-
-  useEffect(() => {
-    if (!isLoaded || !map || !compassRef.current) return
-
-    const compass = compassRef.current
-
-    const updateRotation = () => {
-      const bearing = map.getBearing()
-      const pitch = map.getPitch()
-      compass.style.transform = `rotateX(${pitch}deg) rotateZ(${-bearing}deg)`
-    }
-
-    map.on('rotate', updateRotation)
-    map.on('pitch', updateRotation)
-    updateRotation()
-
-    return () => {
-      map.off('rotate', updateRotation)
-      map.off('pitch', updateRotation)
-    }
-  }, [isLoaded, map])
-
-  return (
-    <ControlButton onClick={onClick} label="Reset bearing to north">
-      <svg
-        ref={compassRef}
-        viewBox="0 0 24 24"
-        className="size-5 transition-transform duration-200"
-        style={{ transformStyle: 'preserve-3d' }}
-      >
-        <path d="M12 2L16 12H12V2Z" className="fill-red-500" />
-        <path d="M12 2L8 12H12V2Z" className="fill-red-300" />
-        <path d="M12 22L16 12H12V22Z" className="fill-muted-foreground/60" />
-        <path d="M12 22L8 12H12V22Z" className="fill-muted-foreground/30" />
-      </svg>
-    </ControlButton>
-  )
-}
-
-type MapPopupProps = {
-  /** Longitude coordinate for popup position */
-  longitude: number
-  /** Latitude coordinate for popup position */
-  latitude: number
-  /** Callback when popup is closed */
-  onClose?: () => void
-  /** Popup content */
-  children: ReactNode
-  /** Additional CSS classes for the popup container */
-  className?: string
-  /** Show a close button in the popup (default: false) */
-  closeButton?: boolean
-} & Omit<PopupOptions, 'className' | 'closeButton'>
-
-function MapPopup({
-  longitude,
-  latitude,
-  onClose,
-  children,
-  className,
-  closeButton = false,
-  ...popupOptions
-}: MapPopupProps) {
-  const { map } = useMap()
-  const popupRef = useRef<MapLibreGL.Popup | null>(null)
-  const popupOptionsRef = useRef(popupOptions)
-  const [container, setContainer] = useState<HTMLDivElement | null>(null)
-
-  // Create container only on client side to avoid SSR hydration mismatch
-  useEffect(() => {
-    setContainer(document.createElement('div'))
-  }, [])
-
-  useEffect(() => {
-    if (!map || !container) return
-
-    const popup = new MapLibreGL.Popup({
-      offset: 16,
-      ...popupOptions,
-      closeButton: false,
-    })
-      .setMaxWidth('none')
-      .setDOMContent(container)
-      .setLngLat([longitude, latitude])
-      .addTo(map)
-
-    const onCloseProp = () => onClose?.()
-
-    popup.on('close', onCloseProp)
-
-    popupRef.current = popup
-
-    return () => {
-      popup.off('close', onCloseProp)
-      if (popup.isOpen()) {
-        popup.remove()
-      }
-      popupRef.current = null
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [map, container])
-
-  useEffect(() => {
-    popupRef.current?.setLngLat([longitude, latitude])
-  }, [longitude, latitude])
-
-  useEffect(() => {
-    if (!popupRef.current) return
-    const prev = popupOptionsRef.current
-
-    if (prev.offset !== popupOptions.offset) {
-      popupRef.current.setOffset(popupOptions.offset ?? 16)
-    }
-    if (prev.maxWidth !== popupOptions.maxWidth && popupOptions.maxWidth) {
-      popupRef.current.setMaxWidth(popupOptions.maxWidth ?? 'none')
-    }
-
-    popupOptionsRef.current = popupOptions
-  }, [popupOptions])
-
-  const handleClose = () => {
-    popupRef.current?.remove()
-    onClose?.()
-  }
-
-  // Don't render until container is available (client-side only)
-  if (!container) return null
-
-  return createPortal(
-    <div
-      className={cn(
-        'relative rounded-md border bg-popover p-3 text-popover-foreground shadow-md animate-in fade-in-0 zoom-in-95',
-        className
-      )}
-    >
-      {closeButton && (
-        <button
-          type="button"
-          onClick={handleClose}
-          className="absolute top-1 right-1 z-10 rounded-sm opacity-70 ring-offset-background transition-opacity hover:opacity-100 focus:outline-none focus:ring-2 focus:ring-ring focus:ring-offset-2"
-          aria-label="Close popup"
-        >
-          <X className="h-4 w-4" />
-          <span className="sr-only">Close</span>
-        </button>
-      )}
-      {children}
-    </div>,
-    container
-  )
-}
-
-type MapRouteProps = {
-  /** Optional unique identifier for the route layer */
-  id?: string
-  /** Array of [longitude, latitude] coordinate pairs defining the route */
-  coordinates: [number, number][]
-  /** Line color as CSS color value (default: theme-aware blue) */
-  color?: string
-  /** Line width in pixels (default: 4) */
-  width?: number
-  /** Line opacity from 0 to 1 (default: 0.85) */
-  opacity?: number
-  /** Dash pattern [dash length, gap length] for dashed lines */
-  dashArray?: [number, number]
-  /** Callback when the route line is clicked */
-  onClick?: () => void
-  /** Callback when mouse enters the route line */
-  onMouseEnter?: () => void
-  /** Callback when mouse leaves the route line */
-  onMouseLeave?: () => void
-  /** Whether the route is interactive (shows pointer cursor on hover). Default: true */
-  interactive?: boolean
-  /** Line color when hovered (default: theme-aware highlight) */
-  hoverColor?: string
-  /** Line width when hovered (default: width + 2) */
-  hoverWidth?: number
-}
-
-// Theme-aware default colors for routes
-const routeColors = {
-  light: {
-    default: '#1d4ed8', // blue-700 - darker blue for better contrast on light maps
-    hover: '#1e40af', // blue-800
-    casing: '#ffffff', // white outline for contrast
-  },
-  dark: {
-    default: '#3b82f6', // blue-500
-    hover: '#60a5fa', // blue-400
-    casing: '#1e293b', // slate-800 outline for contrast
-  },
-}
-
-function MapRoute({
-  id: customId,
-  coordinates,
-  color,
-  width = 4,
-  opacity = 0.85,
-  dashArray,
-  onClick,
-  onMouseEnter,
-  onMouseLeave,
-  interactive = true,
-  hoverColor,
-  hoverWidth,
-}: MapRouteProps) {
-  const { map, isLoaded } = useMap()
-  const { resolvedTheme } = useTheme()
-  const generatedId = useId()
-  const id = customId ?? generatedId
-  const sourceId = `${ROUTE_SOURCE_PREFIX}${id}`
-  const layerId = `${ROUTE_LAYER_PREFIX}${id}`
-  const casingLayerId = `${ROUTE_LAYER_PREFIX}${id}-casing`
-  const [isHovered, setIsHovered] = useState(false)
-
-  // Compute theme-aware colors
-  const themeColors = resolvedTheme === 'dark' ? routeColors.dark : routeColors.light
-  const effectiveColor = color ?? themeColors.default
-  const effectiveHoverColor = hoverColor ?? themeColors.hover
-  const effectiveHoverWidth = hoverWidth ?? width + 2
-  const casingColor = themeColors.casing
-  const casingWidth = width + 4 // Casing is slightly wider than the route
-
-  // Add source and layer on mount
-  useEffect(() => {
-    if (!isLoaded || !map) return
-
-    map.addSource(sourceId, {
-      type: 'geojson',
-      data: {
-        type: 'Feature',
-        properties: {},
-        geometry: { type: 'LineString', coordinates: [] },
-      },
-    })
-
-    // Add casing layer first (underneath) for better contrast
-    map.addLayer({
-      id: casingLayerId,
-      type: 'line',
-      source: sourceId,
-      layout: { 'line-join': 'round', 'line-cap': 'round' },
-      paint: {
-        'line-color': casingColor,
-        'line-width': casingWidth,
-        'line-opacity': opacity * 0.8,
-      },
-    })
-
-    // Add main route layer on top
-    map.addLayer({
-      id: layerId,
-      type: 'line',
-      source: sourceId,
-      layout: { 'line-join': 'round', 'line-cap': 'round' },
-      paint: {
-        'line-color': effectiveColor,
-        'line-width': width,
-        'line-opacity': opacity,
-        ...(dashArray && { 'line-dasharray': dashArray }),
-      },
-    })
-
-    return () => {
-      try {
-        if (map.getLayer(layerId)) map.removeLayer(layerId)
-        if (map.getLayer(casingLayerId)) map.removeLayer(casingLayerId)
-        if (map.getSource(sourceId)) map.removeSource(sourceId)
-      } catch {
-        // ignore
-      }
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [isLoaded, map, sourceId, layerId, casingLayerId])
-
-  // When coordinates change, update the source data
-  useEffect(() => {
-    if (!isLoaded || !map || coordinates.length < 2) return
-
-    const source = map.getSource(sourceId) as MapLibreGL.GeoJSONSource
-    if (source) {
-      source.setData({
-        type: 'Feature',
-        properties: {},
-        geometry: { type: 'LineString', coordinates },
-      })
-    }
-  }, [isLoaded, map, coordinates, sourceId])
-
-  // Update paint properties when props change
-  // We track the "base" values and apply them, but if hovered, re-apply hover styles after
-  useEffect(() => {
-    if (!isLoaded || !map || !map.getLayer(layerId)) return
-
-    // Update casing properties
-    if (map.getLayer(casingLayerId)) {
-      map.setPaintProperty(casingLayerId, 'line-color', casingColor)
-      map.setPaintProperty(
-        casingLayerId,
-        'line-width',
-        (isHovered ? effectiveHoverWidth : width) + 4
-      )
-      map.setPaintProperty(casingLayerId, 'line-opacity', opacity * 0.8)
-    }
-
-    // Always update base properties
-    map.setPaintProperty(layerId, 'line-color', isHovered ? effectiveHoverColor : effectiveColor)
-    map.setPaintProperty(layerId, 'line-width', isHovered ? effectiveHoverWidth : width)
-    map.setPaintProperty(layerId, 'line-opacity', opacity)
-
-    // Handle dashArray - explicitly clear when undefined (fix: dashArray never clears)
-    if (dashArray) {
-      map.setPaintProperty(layerId, 'line-dasharray', dashArray)
-    } else {
-      // Use [1, 0] to force solid line - empty array [] can be invalid in MapLibre spec
-      map.setPaintProperty(layerId, 'line-dasharray', [1, 0])
-    }
-  }, [
-    isLoaded,
-    map,
-    layerId,
-    casingLayerId,
-    effectiveColor,
-    effectiveHoverColor,
-    casingColor,
-    width,
-    effectiveHoverWidth,
-    opacity,
-    dashArray,
-    isHovered,
-  ])
-
-  // Reset hover state when interactive becomes false
-  useEffect(() => {
-    if (!interactive && isHovered) {
-      setIsHovered(false)
-      if (map) {
-        map.getCanvas().style.cursor = ''
-      }
-    }
-  }, [interactive, isHovered, map])
-
-  // Handle hover and click events
-  useEffect(() => {
-    if (!isLoaded || !map) return
-
-    const handleMouseEnter = () => {
-      // When not interactive, don't fire callbacks or change styles
-      if (!interactive) return
-
-      map.getCanvas().style.cursor = 'pointer'
-      // Apply hover styles
-      if (map.getLayer(layerId)) {
-        map.setPaintProperty(layerId, 'line-color', effectiveHoverColor)
-        map.setPaintProperty(layerId, 'line-width', effectiveHoverWidth)
-      }
-      setIsHovered(true)
-      onMouseEnter?.()
-    }
-
-    const handleMouseLeave = () => {
-      // When not interactive, don't fire callbacks or change styles
-      if (!interactive) return
-
-      map.getCanvas().style.cursor = ''
-      // Restore original styles
-      if (map.getLayer(layerId)) {
-        map.setPaintProperty(layerId, 'line-color', effectiveColor)
-        map.setPaintProperty(layerId, 'line-width', width)
-      }
-      setIsHovered(false)
-      onMouseLeave?.()
-    }
-
-    const handleClick = () => {
-      // When not interactive, don't fire click callback
-      if (!interactive) return
-      onClick?.()
-    }
-
-    map.on('mouseenter', layerId, handleMouseEnter)
-    map.on('mouseleave', layerId, handleMouseLeave)
-    map.on('click', layerId, handleClick)
-
-    return () => {
-      map.off('mouseenter', layerId, handleMouseEnter)
-      map.off('mouseleave', layerId, handleMouseLeave)
-      map.off('click', layerId, handleClick)
-    }
-  }, [
-    isLoaded,
-    map,
-    layerId,
-    interactive,
-    onClick,
-    onMouseEnter,
-    onMouseLeave,
-    effectiveColor,
-    effectiveHoverColor,
-    width,
-    effectiveHoverWidth,
-  ])
-
-  return null
-}
-
-type MapClusterLayerProps<P extends GeoJSON.GeoJsonProperties = GeoJSON.GeoJsonProperties> = {
-  /** GeoJSON FeatureCollection data or URL to fetch GeoJSON from */
-  data: string | GeoJSON.FeatureCollection<GeoJSON.Point, P>
-  /** Maximum zoom level to cluster points on (default: 14) */
-  clusterMaxZoom?: number
-  /** Radius of each cluster when clustering points in pixels (default: 50) */
-  clusterRadius?: number
-  /** Colors for cluster circles: [small, medium, large] based on point count (default: ["#51bbd6", "#f1f075", "#f28cb1"]) */
-  clusterColors?: [string, string, string]
-  /** Point count thresholds for color/size steps: [medium, large] (default: [100, 750]) */
-  clusterThresholds?: [number, number]
-  /** Color for unclustered individual points (default: "#3b82f6") */
-  pointColor?: string
-  /** Callback when an unclustered point is clicked */
-  onPointClick?: (feature: GeoJSON.Feature<GeoJSON.Point, P>, coordinates: [number, number]) => void
-  /** Callback when a cluster is clicked. If not provided, zooms into the cluster */
-  onClusterClick?: (clusterId: number, coordinates: [number, number], pointCount: number) => void
-}
-
-function MapClusterLayer<P extends GeoJSON.GeoJsonProperties = GeoJSON.GeoJsonProperties>({
-  data,
-  clusterMaxZoom = 14,
-  clusterRadius = 50,
-  clusterColors = ['#51bbd6', '#f1f075', '#f28cb1'],
-  clusterThresholds = [100, 750],
-  pointColor = '#3b82f6',
-  onPointClick,
-  onClusterClick,
-}: MapClusterLayerProps<P>) {
-  const { map, isLoaded } = useMap()
-  const id = useId()
-  const sourceId = `cluster-source-${id}`
-  const clusterLayerId = `clusters-${id}`
-  const clusterCountLayerId = `cluster-count-${id}`
-  const unclusteredLayerId = `unclustered-point-${id}`
-
-  const stylePropsRef = useRef({
-    clusterColors,
-    clusterThresholds,
-    pointColor,
-  })
-
-  // Add source and layers on mount
-  useEffect(() => {
-    if (!isLoaded || !map) return
-
-    // Add clustered GeoJSON source
-    map.addSource(sourceId, {
-      type: 'geojson',
-      data: typeof data === 'string' ? data : data,
-      cluster: true,
-      clusterMaxZoom,
-      clusterRadius,
-    })
-
-    // Add cluster circles layer
-    map.addLayer({
-      id: clusterLayerId,
-      type: 'circle',
-      source: sourceId,
-      filter: ['has', 'point_count'],
-      paint: {
-        'circle-color': [
-          'step',
-          ['get', 'point_count'],
-          clusterColors[0],
-          clusterThresholds[0],
-          clusterColors[1],
-          clusterThresholds[1],
-          clusterColors[2],
-        ],
-        'circle-radius': [
-          'step',
-          ['get', 'point_count'],
-          20,
-          clusterThresholds[0],
-          30,
-          clusterThresholds[1],
-          40,
-        ],
-      },
-    })
-
-    // Add cluster count text layer
-    map.addLayer({
-      id: clusterCountLayerId,
-      type: 'symbol',
-      source: sourceId,
-      filter: ['has', 'point_count'],
-      layout: {
-        'text-field': '{point_count_abbreviated}',
-        'text-size': 12,
-      },
-      paint: {
-        'text-color': '#fff',
-      },
-    })
-
-    // Add unclustered point layer
-    map.addLayer({
-      id: unclusteredLayerId,
-      type: 'circle',
-      source: sourceId,
-      filter: ['!', ['has', 'point_count']],
-      paint: {
-        'circle-color': pointColor,
-        'circle-radius': 6,
-      },
-    })
-
-    return () => {
-      try {
-        if (map.getLayer(clusterCountLayerId)) map.removeLayer(clusterCountLayerId)
-        if (map.getLayer(unclusteredLayerId)) map.removeLayer(unclusteredLayerId)
-        if (map.getLayer(clusterLayerId)) map.removeLayer(clusterLayerId)
-        if (map.getSource(sourceId)) map.removeSource(sourceId)
-      } catch {
-        // ignore
-      }
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [isLoaded, map, sourceId])
-
-  // Update source data when data prop changes (only for non-URL data )
-  useEffect(() => {
-    if (!isLoaded || !map || typeof data === 'string') return
-
-    const source = map.getSource(sourceId) as MapLibreGL.GeoJSONSource
-    if (source) {
-      source.setData(data)
-    }
-  }, [isLoaded, map, data, sourceId])
-
-  // Update layer styles when props change
-  useEffect(() => {
-    if (!isLoaded || !map) return
-
-    const prev = stylePropsRef.current
-    const colorsChanged =
-      prev.clusterColors !== clusterColors || prev.clusterThresholds !== clusterThresholds
-
-    // Update cluster layer colors and sizes
-    if (map.getLayer(clusterLayerId) && colorsChanged) {
-      map.setPaintProperty(clusterLayerId, 'circle-color', [
-        'step',
-        ['get', 'point_count'],
-        clusterColors[0],
-        clusterThresholds[0],
-        clusterColors[1],
-        clusterThresholds[1],
-        clusterColors[2],
-      ])
-      map.setPaintProperty(clusterLayerId, 'circle-radius', [
-        'step',
-        ['get', 'point_count'],
-        20,
-        clusterThresholds[0],
-        30,
-        clusterThresholds[1],
-        40,
-      ])
-    }
-
-    // Update unclustered point layer color
-    if (map.getLayer(unclusteredLayerId) && prev.pointColor !== pointColor) {
-      map.setPaintProperty(unclusteredLayerId, 'circle-color', pointColor)
-    }
-
-    stylePropsRef.current = { clusterColors, clusterThresholds, pointColor }
-  }, [
-    isLoaded,
-    map,
-    clusterLayerId,
-    unclusteredLayerId,
-    clusterColors,
-    clusterThresholds,
-    pointColor,
-  ])
-
-  // Handle click events
-  useEffect(() => {
-    if (!isLoaded || !map) return
-
-    // Cluster click handler - zoom into cluster
-    const handleClusterClick = async (
-      e: MapLibreGL.MapMouseEvent & {
-        features?: MapLibreGL.MapGeoJSONFeature[]
-      }
-    ) => {
-      const features = map.queryRenderedFeatures(e.point, {
-        layers: [clusterLayerId],
-      })
-      if (!features.length) return
-
-      const feature = features[0]
-      const clusterId = feature.properties?.cluster_id as number
-      const pointCount = feature.properties?.point_count as number
-      const coordinates = (feature.geometry as GeoJSON.Point).coordinates as [number, number]
-
-      if (onClusterClick) {
-        onClusterClick(clusterId, coordinates, pointCount)
-      } else {
-        // Default behavior: zoom to cluster expansion zoom
-        const source = map.getSource(sourceId) as MapLibreGL.GeoJSONSource
-        const zoom = await source.getClusterExpansionZoom(clusterId)
-        map.easeTo({
-          center: coordinates,
-          zoom,
-        })
-      }
-    }
-
-    // Unclustered point click handler
-    const handlePointClick = (
-      e: MapLibreGL.MapMouseEvent & {
-        features?: MapLibreGL.MapGeoJSONFeature[]
-      }
-    ) => {
-      if (!onPointClick || !e.features?.length) return
-
-      const feature = e.features[0]
-      const coordinates = (feature.geometry as GeoJSON.Point).coordinates.slice() as [
-        number,
-        number,
-      ]
-
-      // Handle world copies
-      while (Math.abs(e.lngLat.lng - coordinates[0]) > 180) {
-        coordinates[0] += e.lngLat.lng > coordinates[0] ? 360 : -360
-      }
-
-      onPointClick(feature as unknown as GeoJSON.Feature<GeoJSON.Point, P>, coordinates)
-    }
-
-    // Cursor style handlers
-    const handleMouseEnterCluster = () => {
-      map.getCanvas().style.cursor = 'pointer'
-    }
-    const handleMouseLeaveCluster = () => {
-      map.getCanvas().style.cursor = ''
-    }
-    const handleMouseEnterPoint = () => {
-      if (onPointClick) {
-        map.getCanvas().style.cursor = 'pointer'
-      }
-    }
-    const handleMouseLeavePoint = () => {
-      map.getCanvas().style.cursor = ''
-    }
-
-    map.on('click', clusterLayerId, handleClusterClick)
-    map.on('click', unclusteredLayerId, handlePointClick)
-    map.on('mouseenter', clusterLayerId, handleMouseEnterCluster)
-    map.on('mouseleave', clusterLayerId, handleMouseLeaveCluster)
-    map.on('mouseenter', unclusteredLayerId, handleMouseEnterPoint)
-    map.on('mouseleave', unclusteredLayerId, handleMouseLeavePoint)
-
-    return () => {
-      map.off('click', clusterLayerId, handleClusterClick)
-      map.off('click', unclusteredLayerId, handlePointClick)
-      map.off('mouseenter', clusterLayerId, handleMouseEnterCluster)
-      map.off('mouseleave', clusterLayerId, handleMouseLeaveCluster)
-      map.off('mouseenter', unclusteredLayerId, handleMouseEnterPoint)
-      map.off('mouseleave', unclusteredLayerId, handleMouseLeavePoint)
-    }
-  }, [isLoaded, map, clusterLayerId, unclusteredLayerId, sourceId, onClusterClick, onPointClick])
-
-  return null
-}
-
-export {
-  Map,
-  useMap,
-  MapMarker,
-  MarkerContent,
-  MarkerPopup,
-  MarkerTooltip,
-  MarkerLabel,
-  MapPopup,
-  MapControls,
-  MapRoute,
-  MapClusterLayer,
-}
+export { Map, useMap }
